removing unused code and moving inline code that is...
... used only once in the checks so that the check routines are more evident to the reader.
